# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  originalPrice: number;
  image: string;
  rating: number;
  reviews: number;
  badge: string;
  description: string;
  stock: number;
  featured: boolean;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  originalPrice: number;
  /** @minLength 1 */
  image: string;
  description: string;
  /** @minimum 0 */
  stock: number;
  featured: boolean;
  badge?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  originalPrice?: number;
  /** @minLength 1 */
  image?: string;
  description?: string;
  /** @minimum 0 */
  stock?: number;
  featured?: boolean;
  badge?: string;
}

export interface OrderItem {
  productId: number;
  name: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  price: number;
  image: string;
}

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  location: string;
  items: OrderItem[];
  total: number;
  paymentMethod: string;
  /** @nullable */
  paymentReceipt?: string | null;
  /** @nullable */
  customerImage?: string | null;
  status: string;
  createdAt: string;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  bank_transfer: 'bank_transfer',
  cash_on_delivery: 'cash_on_delivery',
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  location: string;
  /** @minItems 1 */
  items: OrderItem[];
  /** @minimum 0 */
  total: number;
  paymentMethod: OrderInputPaymentMethod;
  paymentReceipt?: string;
  customerImage?: string;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  packed: 'packed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  status: OrderUpdateStatus;
}

export interface AdminSummary {
  totalProducts: number;
  totalOrders: number;
  pendingOrders: number;
  revenue: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};

export type ListOrdersParams = {
status?: string;
};