# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Store catalog operations
  - name: orders
    description: Customer order operations
  - name: admin
    description: Store management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Add a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
  /orders:
    get:
      operationId: listOrders
      tags: [admin]
      summary: List orders for store management
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    patch:
      operationId: updateOrder
      tags: [admin]
      summary: Update an order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get store summary
      responses:
        "200":
          description: Store summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, price, originalPrice, image, rating, reviews, badge, description, stock, featured]
      properties:
        id: { type: number }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        image: { type: string }
        rating: { type: number }
        reviews: { type: number }
        badge: { type: string }
        description: { type: string }
        stock: { type: number }
        featured: { type: boolean }
    ProductInput:
      type: object
      required: [name, category, price, originalPrice, image, description, stock, featured]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        originalPrice: { type: number, minimum: 0 }
        image: { type: string, minLength: 1 }
        description: { type: string }
        stock: { type: number, minimum: 0 }
        featured: { type: boolean }
        badge: { type: string }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        originalPrice: { type: number, minimum: 0 }
        image: { type: string, minLength: 1 }
        description: { type: string }
        stock: { type: number, minimum: 0 }
        featured: { type: boolean }
        badge: { type: string }
    OrderItem:
      type: object
      required: [productId, name, quantity, price, image]
      properties:
        productId: { type: number }
        name: { type: string }
        quantity: { type: number, minimum: 1 }
        price: { type: number, minimum: 0 }
        image: { type: string }
    Order:
      type: object
      required: [id, customerName, email, phone, location, items, total, paymentMethod, status, createdAt]
      properties:
        id: { type: number }
        customerName: { type: string }
        email: { type: string }
        phone: { type: string }
        location: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        total: { type: number }
        paymentMethod: { type: string }
        paymentReceipt: { type: ["string", "null"] }
        customerImage: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    OrderInput:
      type: object
      required: [customerName, email, phone, location, items, total, paymentMethod]
      properties:
        customerName: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        phone: { type: string, minLength: 7 }
        location: { type: string, minLength: 1 }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/OrderItem" }
        total: { type: number, minimum: 0 }
        paymentMethod: { type: string, enum: [bank_transfer, cash_on_delivery] }
        paymentReceipt: { type: string }
        customerImage: { type: string }
    OrderUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, confirmed, packed, shipped, delivered, cancelled] }
    AdminSummary:
      type: object
      required: [totalProducts, totalOrders, pendingOrders, revenue]
      properties:
        totalProducts: { type: number }
        totalOrders: { type: number }
        pendingOrders: { type: number }
        revenue: { type: number }