# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  location: string;
  items: OrderItem[];
  total: number;
  paymentMethod: string;
  /** @nullable */
  paymentReceipt?: string | null;
  /** @nullable */
  customerImage?: string | null;
  status: string;
  createdAt: string;
}