# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  location: string;
  /** @minItems 1 */
  items: OrderItem[];
  /** @minimum 0 */
  total: number;
  paymentMethod: OrderInputPaymentMethod;
  paymentReceipt?: string;
  customerImage?: string;
}