# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  name: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  price: number;
  image: string;
}