# lib/db/src/schema/orders.ts
import { createInsertSchema } from "drizzle-zod";
import { integer, jsonb, pgTable, real, serial, text, timestamp } from "drizzle-orm/pg-core";
import { z } from "zod/v4";

export type StoredOrderItem = {
  productId: number;
  name: string;
  quantity: number;
  price: number;
  image: string;
};

export const ordersTable = pgTable("orders", {
  id: serial("id").primaryKey(),
  customerName: text("customer_name").notNull(),
  email: text("email").notNull(),
  phone: text("phone").notNull(),
  location: text("location").notNull(),
  items: jsonb("items").$type<StoredOrderItem[]>().notNull(),
  total: real("total").notNull(),
  paymentMethod: text("payment_method").notNull(),
  paymentReceipt: text("payment_receipt"),
  customerImage: text("customer_image"),
  status: text("status").notNull().default("pending"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertOrderSchema = createInsertSchema(ordersTable).omit({ id: true, createdAt: true });
export type InsertOrder = z.infer<typeof insertOrderSchema>;
export type Order = typeof ordersTable.$inferSelect;